nmc: Clear pipeline when changing PSWR
If interrupt was pending at the same moment when it was
unmasked in PSWR it will arrive not immediately (e.g arm),
but after a few cycles, enough to chew up a few commands.

This may break some code&tests that rely on interrupt arriving
at the precise moment of exiting the atomic section.

BUT! If we forcefully flush the pipeline, we get behaviour that's
more similar to what we're used to at the expence of a few cycles

Signed-off-by: Andrew Andrianov <andrew@ncrmnt.org>